% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_models.R
\name{final_models}
\alias{final_models}
\title{final models
takes the models functions and train them with differents
seeds each time that calls the training functions. And saves
the models in rds files.}
\usage{
final_models()
}
\description{
final models
takes the models functions and train them with differents
seeds each time that calls the training functions. And saves
the models in rds files.
}
