% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_datasets.R
\name{load_datasets}
\alias{load_datasets}
\title{Datasets Loader}
\usage{
load_datasets(name_dataset)
}
\arguments{
\item{name_dataset:}{character type data set name.
charges a data set saved in the ./datasets folder}
}
\value{
sensors dataset as dataFrame with string as factor
}
\description{
Datasets Loader
}
