% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_classificator.R
\name{train_model_classificator}
\alias{train_model_classificator}
\title{Classificator model training
This takes a sensor shapes dataset to training a model to
classify the shape of a wall taken the data given by ultrasonic
and laser sensors.}
\usage{
train_model_classificator()
}
\description{
Classificator model training
This takes a sensor shapes dataset to training a model to
classify the shape of a wall taken the data given by ultrasonic
and laser sensors.
}
