% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_distance.R
\name{train_model_distance}
\alias{train_model_distance}
\title{Training of distance models
Takes the dataset previously added in the folder datasets,
uses it for training a linear regression model applying the caret
package, the training is set to a random seed each time.
Takes a 80 % from data in training processes and a 20 % for testing}
\usage{
train_model_distance()
}
\description{
Training of distance models
Takes the dataset previously added in the folder datasets,
uses it for training a linear regression model applying the caret
package, the training is set to a random seed each time.
Takes a 80 % from data in training processes and a 20 % for testing
}
